Moves onboard to proper folder
